Split vid module.
